+++ src/extension.ts
@@ -6,7 +6,7 @@ import { registerActiveEditorChangeListener } from "./events";
 import { loadAccessToken } from "./services/githubService";
 
 export function activate(context: vscode.ExtensionContext) {
-  console.log('Congratulations, your extension "hexo-github" is now active!');
+  console.log('Congratulations, your extension "vscode-hexo-github" is now active!');
 
   // 加载本地token
   loadAccessToken();